Profiles RC finished
Profiles should be working by now completely with the ipfixcol
reloading. Needs to be tested in the clean environment (based on this
commit) for bugs.
Folder system was reworked. Installation notes in readme we're updated
but they're not finished. Next commit should be ready to be merged with
the master.